refactor: re-org models in api/public
