+++ api/public/user/model.py
@@ -1,11 +1,11 @@
 from typing import List, Optional, TYPE_CHECKING
 from pydantic import FileUrl, EmailStr
 from sqlmodel import Field, SQLModel, Relationship
-from models import TimeStampMixin
+from api.mixins.timestamp import TimeStampMixin
 
 if TYPE_CHECKING:
-    from models.team import Team
-    from models.post import ChangelogPost
+    from api.public.team.model import Team
+    from api.public.post.model import ChangelogPost
 
 
 class User(TimeStampMixin, SQLModel, table=True):
